Test coverage and some refactoring:
1. Converted the like() calls in t/items-are-0.t to ok(=~) calls, because
the like() confuses the perl debugger: http://xrl.us/tar7 .

2. Started adding tests to test the parsing conditionals of the various 
elements: now testing textinput and skipHours with and without a namespace.

3. Created a _my_in_element function to see if the parser is inside an element.
This is not a method because it is an XML::Parser::Expat object. This was
partially done to avoid testing for namespaces in skipHours that all of its
RSS versions don't support namespaces.



git-svn-id: https://svn.perl.org/modules/XML-RSS/trunk@8616 50811bd7-b8ce-0310-adc1-d9db26280581