Correct GUID 'ispermalink' handling, simplified, and with tests...
This corrects the handling of 'isPermaLink' to adhere to both spec and
the comment above the actual test in code. New test returns '1' if
isPermalink is unset, or set to anything but 'false'.

Used simplified conditional, and updated tests.
Changed missing attribute test to implicit true, removed missing
attribute GUID test to match explicit-true test.